Move converse-profile plugin into folder
